requirements.txt
google.generativeai
dotenv
requests
aiohttp
bs4
phonenumbers
email-validator